# [REVIEW] Use shallow copy for assign operations