Closes #5739 

This PR simply follows the suggestion made by @mrocklin in #5739 .  Note that this simple change has a significant effect on the results of [the cudf merge benchmark](https://github.com/rapidsai/dask-cuda/blob/branch-0.12/dask_cuda/benchmarks/local_cudf_merge.py), which spends a significant amount of time performing a deep copy for a temporary column assignment: 

**AFTER Changes**:
```
Merge benchmark
--------------------------
Chunk-size  | 100000000
Frac-match  | 0.3
Ignore-size | 1.05 MB
Protocol    | ucx
Device(s)   | 0,1,2,3
rmm-pool    | True
tcp         | True
ib          | False
nvlink      | True
==========================
Total time  | 967.77 ms
Total time  | 826.82 ms
Total time  | 844.98 ms
==========================

```
<details>
<summary>Worker-Worker Transfer Rates</summary>

```
(w1,w2)     | 25% 50% 75% (total nbytes)
--------------------------
(00,01)     | 4.10 GB/s 4.26 GB/s 4.76 GB/s (7.20 GB)
(00,02)     | 4.69 GB/s 5.00 GB/s 8.64 GB/s (14.40 GB)
(00,03)     | 4.72 GB/s 5.06 GB/s 5.35 GB/s (7.20 GB)
(01,00)     | 4.09 GB/s 4.64 GB/s 4.84 GB/s (7.20 GB)
(01,02)     | 4.37 GB/s 5.16 GB/s 5.52 GB/s (9.60 GB)
(01,03)     | 4.29 GB/s 4.88 GB/s 5.57 GB/s (12.00 GB)
(02,00)     | 4.46 GB/s 4.87 GB/s 7.50 GB/s (12.00 GB)
(02,01)     | 4.12 GB/s 4.86 GB/s 5.48 GB/s (7.20 GB)
(02,03)     | 4.53 GB/s 4.60 GB/s 5.56 GB/s (7.20 GB)
(03,00)     | 4.31 GB/s 5.36 GB/s 6.17 GB/s (9.60 GB)
(03,01)     | 4.11 GB/s 5.65 GB/s 7.19 GB/s (14.40 GB)
(03,02)     | 4.37 GB/s 5.12 GB/s 5.18 GB/s (7.20 GB)
```
</details>


**BEFORE Changes**:
```
Merge benchmark
--------------------------
Chunk-size  | 100000000
Frac-match  | 0.3
Ignore-size | 1.05 MB
Protocol    | ucx
Device(s)   | 0,1,2,3
rmm-pool    | True
tcp         | True
ib          | False
nvlink      | True
==========================
Total time  | 1.59 s
Total time  | 1.85 s
Total time  | 1.79 s
==========================

```
<details>
<summary>Worker-Worker Transfer Rates</summary>

```
(w1,w2)     | 25% 50% 75% (total nbytes)
--------------------------
(00,01)     | 3.36 GB/s 4.20 GB/s 5.54 GB/s (6.50 GB)
(00,02)     | 4.16 GB/s 4.55 GB/s 4.84 GB/s (8.90 GB)
(00,03)     | 3.97 GB/s 5.10 GB/s 6.93 GB/s (11.30 GB)
(01,00)     | 3.74 GB/s 4.15 GB/s 5.54 GB/s (7.20 GB)
(01,02)     | 3.75 GB/s 4.73 GB/s 6.42 GB/s (10.30 GB)
(01,03)     | 4.30 GB/s 5.22 GB/s 6.56 GB/s (11.00 GB)
(02,00)     | 4.17 GB/s 4.76 GB/s 6.55 GB/s (9.60 GB)
(02,01)     | 4.78 GB/s 5.86 GB/s 6.68 GB/s (12.00 GB)
(02,03)     | 4.92 GB/s 5.76 GB/s 6.14 GB/s (7.20 GB)
(03,00)     | 3.60 GB/s 5.56 GB/s 7.38 GB/s (10.30 GB)
(03,01)     | 3.50 GB/s 4.68 GB/s 5.75 GB/s (9.60 GB)
(03,02)     | 4.39 GB/s 4.86 GB/s 6.59 GB/s (10.30 GB)
```
</details>

**TODO** (Reason for WIP status):
- [x] Address test failures

- [ ] Tests added / passed
- [x] Passes `black dask` / `flake8 dask`