@rjzamora 2019-12-20
**Note**:  It appears that the current implementation is problematic for `test_attribute_assignment`:

```python
import pandas as pd
import dask.dataframe as dd

df = pd.DataFrame({"x": [1, 2, 3, 4, 5], "y": [1.0, 2.0, 3.0, 4.0, 5.0]})
ddf = dd.from_pandas(df, npartitions=2)
print("Before ddf Modified...\n", df.y)

ddf.y = ddf.x + ddf.y
ddf.persist()
print("\nAfter ddf Modified...\n", df.y)
```

Output:
```
Before ddf Modified...
0    1.0
1    2.0
2    3.0
3    4.0
4    5.0
Name: y, dtype: float64

After ddf Modified...
0     2.0
1     4.0
2     6.0
3     8.0
4    10.0
Name: y, dtype: float64
```