@mrocklin 2019-12-21
> Does that issue only show up with attribute assignment or is it present with setitem too? Pandas strongly recommends against assigning to an attribute.

I don't think that we're using setattr internally.  I think that it's more likely that this is what you were referring to earlier, that a shallow copy might not be enough to sufficiently protect the input dataframe.