@TomAugspurger 2020-02-18
@rjzamora here's a potential workaround: only do a shallow copy when we're adding a new column:

```diff
diff --git a/dask/dataframe/methods.py b/dask/dataframe/methods.py
index e3e393093..e09d8297d 100644
--- a/dask/dataframe/methods.py
+++ b/dask/dataframe/methods.py
@@ -263,8 +263,13 @@ def set_column(df, name, value):
 
 
 def assign(df, *pairs):
-    df = df.copy(deep=False)
-    for name, val in dict(partition(2, pairs)).items():
+    pairs = dict(partition(2, pairs))
+    # deep copy when updating an element, to avoid issues
+    # with modifying the original
+    deep = bool(set(pairs) & set(df.columns))
+
+    df = df.copy(deep=bool(deep))
+    for name, val in pairs.items():
         df[name] = val
     return df
```

I think that avoids the behavior failing the test. Can you check if it retains the performance boost? Do people think this is likely to cause issues elsewhere? I'm having trouble reasoning through things just in my head.