@rjzamora 2020-02-18
Thanks @TomAugspurger!  I will give it a try.  You have good timing - I was just about to revisit this :)