@TomAugspurger 2020-02-19
Great to hear. Should we go ahead and merge this in? I think I've convinced myself that this is safe to do.