@mrocklin 2020-02-21
Woo!

On Wed, Feb 19, 2020 at 9:27 AM Tom Augspurger <notifications@github.com>
wrote:

> Merged #5740 <https://github.com/dask/dask/pull/5740> into master.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/dask/dask/pull/5740?email_source=notifications&email_token=AACKZTH3462RAKUEINHL6ADRDVTYFA5CNFSM4J6DOVJ2YY3PNVWWK3TUL52HS4DFWZEXG43VMVCXMZLOORHG65DJMZUWGYLUNFXW5KTDN5WW2ZLOORPWSZGOWXZ2TCI#event-3052644745>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AACKZTGU4GEITLOQV25KEEDRDVTYFANCNFSM4J6DOVJQ>
> .
>